exercise 4.23

+++ osa4/controllers/blogs.js
@@ -80,6 +80,10 @@ blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const user = request.user
   const blog = await Blog.findById(request.params.id)
 
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
   try {
     if (blog.user.toString() === user.id.toString()) {
       await Blog.findByIdAndDelete(request.params.id)
